jan 16 2023

+++ libc/ft_memset.c
@@ -6,22 +6,22 @@
 /*   By: sgluck <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/11 16:15:16 by sgluck            #+#    #+#             */
-/*   Updated: 2023/01/11 18:33:16 by sgluck           ###   ########.fr       */
+/*   Updated: 2023/01/16 16:00:22 by sgluck           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include <stddef.h>
+
 #include "libft.h"
 
 void	*ft_memset(void *str, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	if (!str)
 		return (NULL);
 	i = 0;
-	while(i < len)
+	while (i < len)
 	{
-		*(unsigned char*)(str + i) = (unsigned char) c;
+		*(unsigned char *)(str + i) = (unsigned char) c;
 		i++;
 	}
 	return (str);
